requirements_base.txt
numpy
pandas
tqdm
argparse
torchtext==0.13.0